+++ lizmap/test/test_table_manager.py
@@ -44,7 +44,7 @@ class TestTableManager(unittest.TestCase):
         definitions = FilterByFormDefinitions()
 
         table_manager = TableManager(
-            None, definitions, None, table, None, None, None, None)
+            None, definitions, None, table, None, None, None, None, None)
 
         json = {
             '0': {
@@ -128,7 +128,7 @@ class TestTableManager(unittest.TestCase):
     def test_form_filter_3_7(self):
         """ Test to write to 3.6 format. """
         table_manager = TableManager(
-            None, FilterByFormDefinitions(), None, QTableWidget(), None, None, None, None)
+            None, FilterByFormDefinitions(), None, QTableWidget(), None, None, None, None, None)
 
         json = {
             '0': {
@@ -188,7 +188,7 @@ class TestTableManager(unittest.TestCase):
         definitions = FilterByLoginDefinitions()
 
         table_manager = TableManager(
-            None, definitions, None, table, None, None, None, None)
+            None, definitions, None, table, None, None, None, None, None)
 
         json = {
             'lines': {
@@ -217,7 +217,7 @@ class TestTableManager(unittest.TestCase):
     def test_dataviz_definitions(self):
         """Test dataviz collections keys."""
         table_manager = TableManager(
-            None, DatavizDefinitions(), None, QTableWidget(), None, None, None, None)
+            None, DatavizDefinitions(), None, QTableWidget(), None, None, None, None, None)
         expected = [
             'type', 'title', 'description', 'layerId', 'x_field', 'aggregation',
             'traces', 'html_template', 'layout', 'popup_display_child_plot', 'stacked',
@@ -231,7 +231,7 @@ class TestTableManager(unittest.TestCase):
         definitions = DatavizDefinitions()
 
         table_manager = TableManager(
-            None, definitions, None, table, None, None, None, None)
+            None, definitions, None, table, None, None, None, None, None)
 
         # Lizmap 3.3
         json = {
@@ -284,7 +284,7 @@ class TestTableManager(unittest.TestCase):
         definitions = DatavizDefinitions()
 
         table_manager = TableManager(
-            None, definitions, None, table, None, None, None, None)
+            None, definitions, None, table, None, None, None, None, None)
 
         # Lizmap 3.3
         json = {
@@ -425,7 +425,7 @@ class TestTableManager(unittest.TestCase):
         definitions = DatavizDefinitions()
 
         table_manager = TableManager(
-            None, definitions, None, table, None, None, None, None)
+            None, definitions, None, table, None, None, None, None, None)
 
         # Lizmap 3.3
         json = {
@@ -574,7 +574,7 @@ class TestTableManager(unittest.TestCase):
         definitions = DatavizDefinitions()
 
         table_manager = TableManager(
-            None, definitions, None, table, None, None, None, None)
+            None, definitions, None, table, None, None, None, None, None)
 
         json = {
             '0': {
@@ -633,7 +633,7 @@ class TestTableManager(unittest.TestCase):
     def test_dataviz(self):
         """Test we can read dataviz 3.4 format."""
         table_manager = TableManager(
-            None, DatavizDefinitions(), None, QTableWidget(), None, None, None, None)
+            None, DatavizDefinitions(), None, QTableWidget(), None, None, None, None, None)
 
         json = {
             '0': {
@@ -717,7 +717,7 @@ class TestTableManager(unittest.TestCase):
     def test_filter_by_polygon(self):
         """ Test table manager with filter by polygon. """
         table_manager = TableManager(
-            None, FilterByPolygonDefinitions(), None, QTableWidget(), None, None, None, None)
+            None, FilterByPolygonDefinitions(), None, QTableWidget(), None, None, None, None, None)
 
         json = {
             'config': {
@@ -750,7 +750,7 @@ class TestTableManager(unittest.TestCase):
         definitions = ToolTipDefinitions()
 
         table_manager = TableManager(
-            None, definitions, None, table, None, None, None, None)
+            None, definitions, None, table, None, None, None, None, None)
 
         json = {
             'lines': {
@@ -774,7 +774,7 @@ class TestTableManager(unittest.TestCase):
         definitions = AttributeTableDefinitions()
 
         table_manager = TableManager(
-            None, definitions, None, table, None, None, None, None)
+            None, definitions, None, table, None, None, None, None, None)
 
         json = {
             'lines': {
@@ -803,7 +803,7 @@ class TestTableManager(unittest.TestCase):
         definitions = TimeManagerDefinitions()
 
         table_manager = TableManager(
-            None, definitions, None, table, None, None, None, None)
+            None, definitions, None, table, None, None, None, None, None)
 
         # JSON from 3.3 with all fields
         json = {
@@ -870,7 +870,7 @@ class TestTableManager(unittest.TestCase):
         definitions = EditionDefinitions()
 
         table_manager = TableManager(
-            None, definitions, None, table, None, None, None, None)
+            None, definitions, None, table, None, None, None, None, None)
 
         json = {
             'lines': {
@@ -953,7 +953,7 @@ class TestTableManager(unittest.TestCase):
         definitions = LocateByLayerDefinitions()
 
         table_manager = TableManager(
-            None, definitions, None, table, None, None, None, None)
+            None, definitions, None, table, None, None, None, None, None)
 
         json = {
             'lines': {
@@ -1034,7 +1034,7 @@ class TestTableManager(unittest.TestCase):
         definitions = AtlasDefinitions()
 
         table_manager = TableManager(
-            None, definitions, AtlasEditionDialog, table, None, None, None, None)
+            None, definitions, AtlasEditionDialog, table, None, None, None, None, None)
 
         self.assertEqual(table.columnCount(), len(definitions.layer_config.keys()))
 
@@ -1073,7 +1073,7 @@ class TestTableManager(unittest.TestCase):
         definitions._use_single_row = False
 
         table_manager = TableManager(
-            None, definitions, AtlasEditionDialog, table, None, None, None, None)
+            None, definitions, AtlasEditionDialog, table, None, None, None, None, None)
 
         self.assertEqual(table.columnCount(), len(definitions.layer_config.keys()))
 
@@ -1171,7 +1171,7 @@ class TestTableManager(unittest.TestCase):
         definitions = AtlasDefinitions()
         definitions._use_single_row = False
         table_manager = TableManager(
-            None, definitions, AtlasEditionDialog, table, None, None, None, None)
+            None, definitions, AtlasEditionDialog, table, None, None, None, None, None)
 
         json = {
             'layers': [
@@ -1228,7 +1228,7 @@ class TestTableManager(unittest.TestCase):
         }
 
         table_manager = TableManager(
-            None, definitions, AtlasEditionDialog, table, None, None, None, None)
+            None, definitions, AtlasEditionDialog, table, None, None, None, None, None)
 
         self.assertEqual(table_manager.table.rowCount(), 0)
         table_manager.from_json(json)